requirements.txt
pandas
pandas_gbq
numpy
streamlit
streamlit_option_menu
google-cloud-bigquery
google-cloud-aiplatform
protobuf
matplotlib
mpld3